move value fmtr to own mod
